Add role checks to collections API routes

+++ server/api/collections/[collectionId]/index.delete.ts
@@ -1,5 +1,16 @@
 export default defineEventHandler(async (event) => {
   await requireUserSession(event);
+  if (
+    !(await checkUserAuthority(await getUserSession(event), [
+      "curator",
+      "admin",
+    ]))
+  ) {
+    throw createError({
+      status: 401,
+      statusMessage: "You are not authorized for this action",
+    });
+  }
   const result = await useDatabase().sql<DbExecResult>`
     DELETE FROM collection_creator_relation
     WHERE id = ${getRouterParam(event, "collectionId")}
